+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 1. Added support for NVMe-MI responder, gated on a new `nvme-mi` cargo feature
 
+2. Added support for a PLDM for File Transfer requester, triggered on MCTP
+   address assignment.
 
 ### Changed